Added data mover mode dss-hall that transfers data from a dss node to the new hall wide ffb Lustre file system.
The dss-hall mover runs on a psanaNNN node and a new parameter was needed that indicates
that the src host of a file should be used in the data_migration_ana table instead of
the local host. The output path is /ffbdata which is a link on the data mover hosts
(created by ansible).